Replace arbitrary "data" param on /pool with individual fields
This has gone back and forth a few times, but it does seem like there's a
specific need for a "trackingId" that can be used to correlate the request with
the created pool, and that this is the only piece of extra data that needs to
be written to the chain.

Also adds a "config" parameter to the API - not used currently, but this
gives some future-proofing so it will be accepted and ignored by this version
of the connector if it is needed in the future.

Signed-off-by: Andrew Richardson <andrew.richardson@kaleido.io>