Create event subscriptions on a per-pool basis
The app now maintains one "base" subscription for all TokenCreate events, but
allows clients to choose which pools to monitor via a new "/activatepool" API,
which will create a new subscription for all TokenCreate and TransferSingle
events in that pool only. Events are filtered based on the subscription name,
so that events from other pools are silently ignored by each pool subscription.

This comes with two breaking changes:
  - The former "/pool" API has been renamed "/createpool"
  - No transfer events will be delivered until "/activatepool" is called once

Signed-off-by: Andrew Richardson <andrew.richardson@kaleido.io>