Add an "id" to each transfer event which is guaranteed to be unique
Clients need a key that is guaranteed to be unique in order to avoid double-processing
of events. Previously we were using transactionHash, but this will not be unique in
many cases with loops that emit multiple events - for instance:
  - Calling mintNonFungible triggers a loop in Solidity
  - Calling safeBatchTransferFrom triggers a loop in the tokens service

Therefore we'll build a new id based on block number, transaction index, log index,
and an optional application-specified index. This will reliably be unique for every
event.

Signed-off-by: Andrew Richardson <andrew.richardson@kaleido.io>