Add namespace to all subscriptions and events
The /init and /activatepool APIs now accept a namespace, which will
be incorporated into the subscription names on ethconnect, and will
be unpacked and returned on all websocket events.

The field is mandatory on /activatepool (therefore requires a new
version of FireFly) but optional on /init (so older CLIs that attempt
/init will not be broken). Newer FireFly versions should also perform
/init once for each namespace utilizing this token connector.

Going forward, all ethconnect subscription names  will have a prefix
of "fft:" so they can more easily be recognized. This prefix can also
be versioned in the future if further subscription changes are needed.

Existing subscriptions will continue to function, but to avoid issues
when adding new subscriptions, the following migration steps are highly
recommended after upgrading the token connector:
- Stop the token connector and FireFly
- On ethconnect, remove all token subscriptions not starting with "fft:"
- Start the token connector and FireFly, and verify new subscriptions are
  created with the "fft:" prefix

Signed-off-by: Andrew Richardson <andrew.richardson@kaleido.io>