Drop narrow "poolId" formatting in favor of "startId/endId"
Pools on ERC1155 are really just ranges of the token ID space. Going forward,
all pools will be expressed with a startId and endId. This makes the service
logic much less bound to the specific decisions made in the sample contract
around partitioning it in chunks of exactly 2^128.

The old poolId format such as "F1" or "N2" will still be parsed as expected,
but will not be attached to any new pools.

Signed-off-by: Andrew Richardson <andrew.richardson@kaleido.io>