Extract a clearer IERC1155Factory interface
Split the concept of a "pool factory" (which encompasses a single method and
a single event) away from all of the other opinionated spellings in the
IERC1155MixedFungible interface.

The parameters of the TokenPoolCreation event have also been altered, but
both the old and new parameters will be parsed by the listener.

Signed-off-by: Andrew Richardson <andrew.richardson@kaleido.io>