Extend ERC1155URIStorage
This makes the contracts more consistent with OpenZeppelin examples. No
signatures are changed, but there is one slight change in behavior: if
you specify a "base URI", then individual token URIs are always
understood to be a suffix for that base URI. The previous contract
would ignore the base URI when a token URI was specified.

Signed-off-by: Andrew Richardson <andrew.richardson@kaleido.io>